requirements.txt
Django==1.9.1
django-filter==0.12.0
django-markdown==0.8.4
django-pipeline==1.6.4
djangorestframework==3.3.2
psycopg2==2.6.1
